Refactor & simplification
This work was done in order to prepare getting this repo ready to be a PyPi package - a one-liner you can call the class of from your main method rather than having actual logic in main.py. Now main.py is simply a recommendation/guide rather than something that needs to be on your Pi Pico.